Remove old test module
